requirements.txt
langchain
langchain-openai  # Or langchain-groq for Grok
langgraph
requests
google-cloud-functions
google-cloud-scheduler
google-cloud-secret-manager
google-cloud-firestore
python-dotenv
